Body and Origin

Index

The Body and Origin types

ConstrainedDynamics.Body — Type

mutable struct Body{T} <: ConstrainedDynamics.AbstractBody{T}

A Body is described by a mass m and inertia J. For referencing, each Body is assigned a unique id when added to a Mechanism. A name can also be set. The state of a Body contains position, orientation, and velocity information (stored as a State). For Visualization, a shape can be assigned to a Body or a Body can be directly created from a Shape.

ConstrainedDynamics.Origin — Type

mutable struct Origin{T} <: ConstrainedDynamics.AbstractBody{T}

The Origin is the root of a mechanism. For referencing, the Origin is assigned a unique id when added to a Mechanism. A name can also be set. For Visualization, a shape can be assigned to the Origin.

Functions