Constraints

EqualityConstraint

ConstrainedDynamics.EqualityConstraint — Type

mutable struct EqualityConstraint{T, N, Nc, Cs} <: ConstrainedDynamics.AbstractConstraint{T, N}

An EqualityConstraint is a component of a Mechanism and is used to describe the kinematic relation between two or more Bodys. Typically, an EqualityConstraint should not be created directly. Use the joint prototypes instead, for example:

EqualityConstraint(Revolute(body1, body2, rotation_axis)).

Important attributes

id: The unique ID of a constraint. Assigned when added to a Mechanism.

name: The name of a constraint. The name is taken from a URDF or can be assigned by the user.

parentid: The ID of the parent body.

childids: The IDs of the child bodies.

Joint Prototypes

The functions below can be used to create joints between two bodies. Note that body1 can also be the origin. The following arguments are available for most joints:

axis: The rotation axis or plane axis.

p1: The joint connection point for body1 in body1's frame

p2: The joint connection point for body2 in body2's frame

qoffsett: The orientation offset of body2 relative to body1

ConstrainedDynamics.Cylindrical — Method

Cylindrical(body1, body2, axis; p1, p2, qoffset)

A cylindrical joint between two bodies.

ConstrainedDynamics.CylindricalFree — Method

CylindricalFree(body1, body2, axis; p1, p2)

A cylindrical joint between two bodies with unconstrained orientation (point-on-line).

ConstrainedDynamics.Fixed — Method

Fixed(body1::AbstractBody, body2; p1, p2, qoffset)

A fixed connection between two bodies.

ConstrainedDynamics.FixedOrientation — Method

FixedOrientation(body1, body2; qoffset)

Fixed orientation between two bodies (chicken's head).

ConstrainedDynamics.Floating — Method

Floating(body1, body2)

An unconstrained connection between two bodies (connection between floating base and origin).

ConstrainedDynamics.Planar — Method

Planar(body1, body2, axis; p1, p2, qoffset)

A planar joint between two bodies.

ConstrainedDynamics.PlanarAxis — Method

PlanarAxis(body1, body2, axis; p1, p2, qoffset)

A planar joint between two bodies with a rotation axis perpendicular to the plane (turtle bot).

ConstrainedDynamics.PlanarFree — Method

PlanarFree(body1, body2, axis; p1, p2)

A planar joint between two bodies with unconstrained orientation.

ConstrainedDynamics.Prismatic — Method

Prismatic(body1, body2, axis; p1, p2, qoffset)

A prismatic joint between two bodies.

ConstrainedDynamics.Revolute — Method

Revolute(body1, body2, axis; p1, p2, qoffset)

A revolute joint between two bodies (pin, continuous, hinge joint).

ConstrainedDynamics.Spherical — Method

Spherical(body1, body2; p1, p2)

A spherical joint between two bodies (ball-and-socket joint).