EqualityConstraint

Index

The EqualityConstraint type

ConstrainedDynamics.EqualityConstraint — Type

mutable struct EqualityConstraint{T, N, Nc, Cs} <: ConstrainedDynamics.AbstractConstraint{T,N}

An EqualityConstraint is used to describe the kinematic relation between two or more Bodys. For referencing, each EqualityConstraint is assigned a unique id when added to a Mechanism. A name can also be set. Typically, an EqualityConstraint should not be created directly, but as a joint (see Joint).

Functions